Do not serialize SOURCE-retained annotations
Also, fix the value of "hasAnnotations" flag to reflect if there are any
_non-source_ annotations on a declaration.

Unfortunately, after this change
IncrementalJsCompilerRunnerTestGenerated$PureKotlin.testAnnotations
starts to fail because of the following problem. The problem is that
annotations on property accessors are not serialized yet on JS (see
SerializerExtensionBase.serializeProperty), yet property proto message
has setterFlags field which has the hasAnnotations flag. Upon the full
rebuild of the code in that test, we correctly write hasAnnotations =
true, but annotations themselves are not serialized. After an
incremental build, we deserialize property setter descriptor, observe
its Annotations object which happens to be an instance of
NonEmptyDeserializedAnnotationsWithPossibleTargets. Now, because
annotations itself are not serialized, that Annotations object has no
annotations, yet its isEmpty always returns false (see the code).
Everything worked correctly before the change because in
DescriptorSerializer.hasAnnotations, we used Annotations.isEmpty and the
result was the same in the full rebuild and in the incremental scenario.
But now we're actually loading annotations, to determine their
retention, and that's why the setterFlags are becoming different here
and the test fails

 #KT-23360 Fixed